Moved AdminConfig into servlet package, renamed to ServletConfig
Things that impact the entire application have been moved into
the 'servlet' package and appropriately renamed.
