+++ kangaroo-servlet-admin/src/main/java/net/krotscheck/kangaroo/servlet/admin/v1/AdminV1API.java
@@ -23,9 +23,9 @@ import net.krotscheck.kangaroo.common.exception.ExceptionFeature;
 import net.krotscheck.kangaroo.common.jackson.JacksonFeature;
 import net.krotscheck.kangaroo.common.version.VersionFeature;
 import net.krotscheck.kangaroo.database.DatabaseFeature;
-import net.krotscheck.kangaroo.servlet.admin.v1.config.AdminConfigurationFactory;
 import net.krotscheck.kangaroo.servlet.admin.v1.filter.OAuth2AuthorizationFilter;
 import net.krotscheck.kangaroo.servlet.admin.v1.resource.UserService;
+import net.krotscheck.kangaroo.servlet.admin.v1.servlet.ServletConfigFactory;
 import org.glassfish.jersey.CommonProperties;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.server.filter.RolesAllowedDynamicFeature;
@@ -56,7 +56,7 @@ public final class AdminV1API extends ResourceConfig {
         register(VersionFeature.class);          // Version response attachment.
 
         // Internal components
-        register(new AdminConfigurationFactory.Binder());
+        register(new ServletConfigFactory.Binder());
 
         // API Authorization
         register(new OAuth2AuthorizationFilter.Binder());
